+++ backend/moyeo/src/main/java/com/mo/moyeo/domain/transaction/transfer/dto/TransferRequest.java
@@ -7,6 +7,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 @Builder
@@ -23,6 +25,6 @@ public class TransferRequest {
 
     @NotNull(message = "금액은 필수 입력값입니다.")
     @Positive(message = "금액은 0보다 커야 합니다.")
-    private Double amount;
+    private BigDecimal amount;
 
 }
